Inclusão da correção na tela do listão fechamento (Diário de Classe > Listão > Fechamento)
