+++ src/SME.SGP.Aplicacao/Queries/Frequencia/ExigeAprovacaoDeNota/ExigeAprovacaoDeNotaQuery.cs
@@ -1,0 +1,15 @@
+﻿using MediatR;
+using SME.SGP.Dominio;
+
+namespace SME.SGP.Aplicacao
+{
+    public class ExigeAprovacaoDeNotaQuery : IRequest<bool>
+    {
+        public Turma Turma { get; set; }
+
+        public ExigeAprovacaoDeNotaQuery(Turma turma)
+        {
+            Turma = turma;
+        }
+    }
+}
